q4_content_creation_assistant/requirements.txt
# Core Framework
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10

# GUI Framework
tkinter-tooltip==2.1.0
customtkinter==5.2.0

# OpenAI API
openai==1.10.0

# Medium API
medium-api==0.6.0

# File operations and utilities
python-dotenv==1.0.0
requests==2.31.0
markdown==3.5.1

# Date and time handling
python-dateutil==2.8.2

# JSON handling
jsonschema==4.20.0

# Async support
aiofiles==23.2.1 